requirements.txt
altgraph==0.17
asn1crypto==1.3.0
azure-common==1.1.25
azure-core==1.7.0
azure-storage-blob==12.3.2
boto3==1.14.22
botocore==1.17.22
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
docutils==0.15.2
idna==2.9
isodate==0.6.0
jmespath==0.10.0
macholib==1.14
modulegraph==0.18
msrest==0.6.17
numpy==1.19.0
oauthlib==3.1.0
oscrypto==1.2.0
pandas==1.0.5
py2app==0.13
pycparser==2.20
pycryptodomex==3.9.8
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.2.9
urllib3==1.25.9
